requirements.txt
torch
nltk
numpy==1.24.3
python-dotenv==1.0.0
streamlit==1.24.0
streamlit_extras
openai==0.27.6
faiss-cpu==1.7.4
streamlit-extras
streamlit-folium
scikit-learn
pandas 
numpy==1.24.3
langchain==0.0.184
PyPDF2==3.0.1
altair==4
tiktoken==0.4.0

# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2
